pyproject.toml
[project]
name = "ai-data-generator"
version = "0.1.0"
description = ""
authors = [
    {name = "Kamil",email = "kkukielka@griddynamics.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "google-genai (>=1.46.0,<2.0.0)",
    "streamlit (>=1.50.0,<2.0.0)",
    "sqlalchemy (>=2.0.0,<3.0.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "guardrails-ai (>=0.6.0,<0.7.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false
